release: prepare v1.0.0
Resumen de cambios principales para la versión 1.0.0:
- Booking: eliminación del campo 'date' en la solicitud (el servidor fija la fecha a hoy). Actualizados DTOs, controlador y servicio (createBookingToday + validaciones de pasado y conflictos).
- Observabilidad: añadido requestId (header X-Request-Id y campo en ApiResponse) con RequestContext + RequestIdFilter.
- Manejo de errores: consolidación en ApiExceptionHandler; incorporación de DomainValidationException para reglas de negocio (422) y mensajes estandarizados sin punto final.
- Privacidad: PaymentMethod ahora persiste solo los últimos 4 dígitos de la tarjeta.
- Repositorios Spring Data JPA formalizados para entidades (Customer, Trainer, Booking, IdentityLink, ProgressRecord, Routine, AttendanceRecord).
- DTOs: ajuste de BookingCreateRequest, BookingResponse y mejoras de nombres/normalización.
- HealthController: nuevo endpoint /health para verificación rápida + requestId.
- Script smoke: robustecido (34 escenarios) con selección dinámica de horarios evitando colisiones (concurrency, delete, booking válido). Pass rate 100%.
- Seed de desarrollo: nuevo data.sql idempotente para baseline trainer/customer/identity y limpieza inicial de bookings.
- README: actualizado con seed, checklist v1.0, formato de respuesta, taxonomía de códigos y eliminación de emojis.
- .gitignore: ignorar archivos H2 (smartgymdb.mv.db, smartgymdb.trace.db) y patrones *.mv.db, *.trace.db.
- Ajustes menores en entidades (normalización de emails, restricciones únicas, índices, embeddables).
- application.yml: confirmación de perfil H2 y parámetros para persistencia.

Listo para etiquetar v1.0.0.

+++ src/main/java/com/smartgym/api/dto/BookingResponse.java
@@ -1,5 +1,5 @@
 package com.smartgym.api.dto;
 
 public record BookingResponse(
-        int id, String customerEmail, String trainerEmail, String date, String time, String note
+        Long id, String customerEmail, String trainerEmail, String date, String time, String note
 ) {}
